(#29) Added `tryToGet()`, adjusted to use API 23 if available.
Upon further review, it made no sense to have the class try to use its own way of obtaining manager instances if the API 23+ way was available. Now you can continue to use the `Managers` class for any API level, and don't have to switch away from it if the target API is 23+. Also there's now a `get()` (which throws exceptions from the standard method) and `tryToGet()` (which catches and logs them, returning null instead).