(#26) Created a database! But is it correct...?
The unit test `testDatabaseCreation()` needs to find a way to peer into the SQLite database's schema after it is created, and verify that the tables were created, and that their columns adhere to the appropriate schema.

This might be simpler as a unit test of the `getTableCreationSQL()` method itself, with the `testDatabaseCreation()` method simply verifying that the database file itself was created.